feat: implement removing cart item
